requirements.txt
# Core dependencies
python-dotenv==1.0.0

# Data manipulation
pandas==2.1.4
numpy==1.26.2

# Async operations
aiohttp==3.9.1
asyncio==3.4.3
websockets==12.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Market data APIs
yfinance==0.2.33
alpaca-py==0.21.0

# LLM Integration
anthropic==0.18.1
openai==1.7.2

# Web framework (for dashboard/API)
fastapi==0.109.0
uvicorn==0.25.0

# Task scheduling
apscheduler==3.10.4

# Configuration
pyyaml==6.0.1

# Logging
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0
